Second Commit
Utilises a for loop to check all the character elements in the first string and maps the characters to their corresponding position in the second string if the HashMap doesn't contain the key. If the HashMap contains the key and the value doesn't match the character in the second string, returns false and exits. Returns true if one type of character in the first string doesn't map to two different characters corresponding in the second string.

Runs in O(n) time complexity based on the length of the first string.

Runs in O(n) space complexity based on how many elements are added into the HashMap.